Replace the fancy quotes with a normal quote character
Some books use fancy quotes that somehow point towards the text they're enclosing (like ゛and another pointing to the left), but they don't seem to render correctly outside of the ebook reader. I had previously replaced them with 《》 but that doesn't actually work now that it's globally replacing them (in a previous un-committed fix I had it just replace one character each and it looked fine, but if two things were quoted it wouldn't do the second one), as they are not separate characters for either side of the text. With the previously commit I had made it globally replace each quote character, but I now realize it's the same character, the font rendering just made it look different. In practice this would make stuff 《quoted like this《. Obviously it's wrong to have two opening quotes! This just replaces both with a normal double quote character. It'll be more accurate to the source usage anyway.